+++ src/forecast.cc
@@ -2,6 +2,8 @@
 
 #include "dailyforecastbox.h"
 #include "events.h"
+#include "pager.h"
+#include <memory>
 
 namespace taranis {
 
@@ -9,22 +11,35 @@ ForecastStack::ForecastStack(int pos_x, int pos_y, int width, int height,
                              std::shared_ptr<Model> model,
                              std::shared_ptr<Icons> icons,
                              std::shared_ptr<Fonts> fonts)
-    : Stack{}, model{model} {
+    : Widget{pos_x, pos_y, width, height}, model{model},
+      pager{std::make_shared<DotPager>(
+          pos_x, pos_y + height - ForecastStack::pager_height, width,
+          ForecastStack::pager_height, std::shared_ptr<Paginated>(this))},
+      stack{std::make_shared<Stack>()} {
+  const auto stack_height =
+      height - ForecastStack::pager_height - ForecastStack::vertical_padding;
+
   this->hourly_forecast_box = std::make_shared<HourlyForecastBox>(
-      pos_x, pos_y, width, height, model, icons, fonts);
+      pos_x, pos_y, width, stack_height, model, icons, fonts);
 
   auto daily_forecast_box = std::make_shared<DailyForecastBox>(
-      pos_x, pos_y, width, height, model, icons, fonts);
+      pos_x, pos_y, width, stack_height, model, icons, fonts);
 
-  this->add("hourly-forecast", this->hourly_forecast_box);
-  this->add("daily-forecast", daily_forecast_box);
+  this->stack->add("hourly-forecast", this->hourly_forecast_box);
+  this->stack->add("daily-forecast", daily_forecast_box);
+
+  const auto pager_update_callback = [=]() {
+    this->pager->paint_and_update_screen();
+  };
+  this->hourly_forecast_box->set_page_changed_callback(pager_update_callback);
+  this->stack->set_current_widget_callback(pager_update_callback);
 
   this->swipe_detector.set_bounding_box(this->get_bounding_box());
 
   auto forecast_to_display = this->model->display_daily_forecast
                                  ? "daily-forecast"
                                  : "hourly-forecast";
-  this->set_current_widget_maybe(forecast_to_display);
+  this->stack->set_current_widget_maybe(forecast_to_display);
 }
 
 int ForecastStack::handle_pointer_event(int event_type, int pointer_pos_x,
@@ -32,7 +47,8 @@ int ForecastStack::handle_pointer_event(int event_type, int pointer_pos_x,
   if (this->handle_possible_swipe(event_type, pointer_pos_x, pointer_pos_y)) {
     return 1;
   }
-  return Stack::handle_pointer_event(event_type, pointer_pos_x, pointer_pos_y);
+  return this->stack->handle_pointer_event(event_type, pointer_pos_x,
+                                           pointer_pos_y);
 }
 
 bool ForecastStack::handle_possible_swipe(int event_type, int pointer_pos_x,
@@ -40,7 +56,7 @@ bool ForecastStack::handle_possible_swipe(int event_type, int pointer_pos_x,
   const auto swipe = this->swipe_detector.guess_event_swipe_type(
       event_type, pointer_pos_x, pointer_pos_y);
   if (swipe != SwipeType::no_swipe) {
-    const auto current_forecast_name = this->current_widget_name();
+    const auto current_forecast_name = this->stack->current_widget_name();
     if (current_forecast_name == "daily-forecast") {
       if (swipe == SwipeType::left_swipe) {
         this->hourly_forecast_box->set_min_forecast_offset();
@@ -78,7 +94,7 @@ bool ForecastStack::handle_key_press(int key) {
 }
 
 bool ForecastStack::handle_key_repeat(int key) {
-  const auto current_forecast_name = this->current_widget_name();
+  const auto current_forecast_name = this->stack->current_widget_name();
   if (current_forecast_name == "hourly-forecast") {
     this->prepare_child_switch_by_key(key);
     this->request_forecast_switch();
@@ -87,7 +103,7 @@ bool ForecastStack::handle_key_repeat(int key) {
 }
 
 bool ForecastStack::handle_key_release(int key) {
-  const auto current_forecast_name = this->current_widget_name();
+  const auto current_forecast_name = this->stack->current_widget_name();
   if (current_forecast_name == "hourly-forecast") {
     if ((key == IV_KEY_PREV and
          this->hourly_forecast_box->get_forecast_offset() ==
@@ -110,8 +126,25 @@ bool ForecastStack::handle_key_release(int key) {
   ;
 }
 
+size_t ForecastStack::page_count() const {
+  return 1 + this->hourly_forecast_box->page_count();
+}
+
+size_t ForecastStack::current_page() const {
+  const auto current_forecast_name = this->stack->current_widget_name();
+  if (current_forecast_name == "hourly-forecast") {
+    return this->hourly_forecast_box->current_page() + 1;
+  }
+  return 0;
+}
+
+void ForecastStack::do_paint() {
+  this->stack->do_paint();
+  this->pager->do_paint();
+}
+
 void ForecastStack::prepare_child_switch_by_key(int key) {
-  const auto current_forecast_name = this->current_widget_name();
+  const auto current_forecast_name = this->stack->current_widget_name();
   if (current_forecast_name == "daily-forecast") {
     if (key == IV_KEY_PREV) {
       this->hourly_forecast_box->set_max_forecast_offset();
@@ -133,7 +166,6 @@ void ForecastStack::switch_forecast() {
   auto forecast_to_display = this->model->display_daily_forecast
                                  ? "daily-forecast"
                                  : "hourly-forecast";
-  this->set_current_widget(forecast_to_display);
+  this->stack->set_current_widget(forecast_to_display);
 }
-
 } // namespace taranis