+++ src/stack.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <functional>
 #include <map>
 #include <memory>
 
@@ -8,6 +9,8 @@
 
 namespace taranis {
 struct Stack : Widget {
+  typedef std::function<void()> CurrentWidgetChangedCallback;
+
   Stack();
 
   void add(const std::string &name, std::shared_ptr<Widget> child);
@@ -19,6 +22,8 @@ struct Stack : Widget {
   void set_current_widget(const std::string &name);
 
   void set_current_widget(const std::shared_ptr<Widget> &widget);
+
+  void set_current_widget_callback(CurrentWidgetChangedCallback callback);
 
   bool is_enabled() const override;
 
@@ -33,17 +38,20 @@ struct Stack : Widget {
 
   bool handle_key_release(int key) override;
 
-protected:
   bool set_current_widget_maybe(const std::string &name);
   // Doesn't paint, use in constructors to set initial state
 
 private:
   typedef std::map<std::string, std::shared_ptr<Widget>> Children;
 
+  CurrentWidgetChangedCallback current_widget_changed_callback;
+
   Children children;
 
   Children::const_iterator current;
 
   std::shared_ptr<KeyEventConsumer> last_key_event_consumer{nullptr};
+
+  void on_current_widget_changed();
 };
 } // namespace taranis